requirements.txt
uvicorn
transformers
lxml
newspaper3k
accelerate
pydub
ffmpeg
ttsmms
fastapi
nltk
sacremoses
pandas
regex
mock
mosestokenizer
bitsandbytes
scipy
sentencepiece
lxml_html_clean
pydantic

